show scheduled times in calendar
